@@ -3,43 +3,39 @@ import 'dart:typed_data';
 
 import 'package:dio/dio.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:fpdart/fpdart.dart';
 import 'package:myapp/core/error/failure.dart';
 import 'package:myapp/core/services/api_service.dart';
-import 'package:myapp/core/utils.dart';
+import 'package:myapp/core/services/path_service.dart';
 
 abstract class ISubLevelAPI {
-  FutureEither<Uint8List?> getZipData(String levelId, int zipId);
+  Future<Uint8List?> getVideo(String levelId, String videoFilename);
 }
 
 class SubLevelAPI implements ISubLevelAPI {
-  SubLevelAPI(this.apiService);
+  SubLevelAPI(this.apiService, this.pathService);
 
   final ApiService apiService;
+  final PathService pathService;
 
   @override
-  FutureEither<Uint8List?> getZipData(
-    String levelId,
-    int zipNum,
-  ) async {
+  Future<Uint8List?> getVideo(String levelId, String videoFilename) async {
     try {
-      final response = await apiService.getCloudStorageData<Uint8List?>(
-        params: ApiParams(
-          endpoint: getLevelZipPath(levelId, zipNum),
-          baseUrl: BaseUrl.s3,
-          method: ApiMethod.get,
-          responseType: ResponseType.bytes,
-        ),
+      final response = await apiService.getCloudStorageData(
+        endpoint: pathService.videoPath(levelId, videoFilename),
+        responseType: ResponseType.bytes,
       );
 
-      return Right(response?.data);
+      return response?.data;
     } on DioException catch (e) {
-      developer.log('Error in SubLevelAPI.getZipData: $e');
-      return Left(Failure(message: e.toString()));
+      developer.log('Error in SubLevelAPI.getVideo: $e');
+      throw Failure(message: e.toString());
     }
   }
 }
 
 final subLevelAPIProvider = Provider<ISubLevelAPI>((ref) {
-  return SubLevelAPI(ref.read(apiServiceProvider));
+  final apiService = ref.read(apiServiceProvider);
+  final pathService = ref.read(pathServiceProvider);
+
+  return SubLevelAPI(apiService, pathService);
 });
