@@ -3,11 +3,5 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 
 final googleSignInProvider = Provider<GoogleSignIn>((ref) {
-  return GoogleSignIn(
-    scopes: [
-      'email',
-      'profile',
-    ],
-    serverClientId: dotenv.env['GOOGLE_SERVER_ID'],
-  );
+  return GoogleSignIn(scopes: ['email', 'profile'], serverClientId: dotenv.env['GOOGLE_SERVER_ID']);
 });
