@@ -2,60 +2,53 @@ import 'package:android_play_install_referrer/android_play_install_referrer.dart
 import 'package:app_links/app_links.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:fpdart/fpdart.dart';
 import 'package:go_router/go_router.dart';
 import 'package:myapp/apis/initialize_api.dart';
+import 'package:myapp/constants/constants.dart';
 import 'package:myapp/core/router/router.dart';
 import 'package:myapp/core/services/file_service.dart';
 import 'package:myapp/core/services/info_service.dart';
 import 'package:myapp/core/shared_pref.dart';
 import 'package:myapp/core/util_types/progress.dart';
-import 'package:myapp/features/sublevel/ordered_ids_notifier.dart';
+import 'package:myapp/features/sublevel/level_controller.dart';
 import 'package:myapp/features/user/user_controller.dart';
 import 'dart:developer' as developer;
-
-import 'package:path_provider/path_provider.dart';
 
 // Global key for navigation
 final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
 
 class InitializeService {
-  UserController userController;
+  UserControllerState userControllerState;
   InitializeAPI initializeAPI;
-  OrderedIdsNotifier orderedIdNotifier;
+  LevelController levelController;
+  UserController userController;
 
   InitializeService({
-    required this.userController,
+    required this.userControllerState,
     required this.initializeAPI,
-    required this.orderedIdNotifier,
+    required this.levelController,
+    required this.userController,
   });
-
-  Future<void> clearAppCache() async {
-    try {
-      final cacheDir = await getTemporaryDirectory();
-      if (cacheDir.existsSync()) {
-        cacheDir.deleteSync(recursive: true);
-      }
-    } catch (e) {
-      developer.log('Unable to kill cache duo to $e');
-    }
-  }
 
   Future<void> initialize() async {
     try {
-      await handleDeepLinking();
-      await storeCyId();
-      await FileService.instance.init();
-      await InfoService.instance.init();
-      await SharedPref.init();
-      await clearAppCache();
-      await orderedIdNotifier.getOrderedIds();
+      // order matters
+      await SharedPref.init(); // first init shared pref
+      await InfoService.instance.init(); // then init info service
+      await initialApiCall(); // then call api because it depends on info service
 
-      await initializeVersion();
+      await Future.wait([
+        storeCyId(), // depend on user
+        FileService.instance.init(),
+        levelController.getOrderedIds(),
+        handleDeepLinking(), // deep linking depends on user
+      ]);
 
-      final currProgress = await SharedPref.getValue(PrefKey.currProgress);
-      final apiUser = await userController.getCurrentUser();
+      final currProgress = SharedPref.get(
+        PrefKey.currProgress(userEmail: userControllerState.currentUser?.email),
+      );
 
+      final apiUser = userControllerState.currentUser;
       if (currProgress == null && apiUser == null) return;
 
       final localLastModified = currProgress?.modified ?? 0;
@@ -65,37 +58,37 @@ class InitializeService {
       localLastModified > apiLastModified
           ? await userController.sync(currProgress!.levelId!, currProgress.subLevel!)
           : await SharedPref.copyWith(
-              PrefKey.currProgress,
-              Progress(
-                level: apiUser?.level,
-                levelId: apiUser?.levelId,
-                maxLevel: apiUser?.maxLevel,
-                maxSubLevel: apiUser?.maxSubLevel,
-                subLevel: apiUser?.subLevel,
-                modified: apiLastModified,
-              ),
-            );
+            PrefKey.currProgress(userEmail: apiUser?.email),
+            Progress(
+              level: apiUser?.level,
+              levelId: apiUser?.levelId,
+              maxLevel: apiUser?.maxLevel,
+              maxSubLevel: apiUser?.maxSubLevel,
+              subLevel: apiUser?.subLevel,
+            ),
+          );
 
-      await SharedPref.storeValue(PrefKey.isFirstLaunch, false);
+      await SharedPref.store(PrefKey.isFirstLaunch, false);
     } catch (e, stackTrace) {
       developer.log('Error during initialize', error: e.toString(), stackTrace: stackTrace);
     }
   }
 
   Future<void> storeCyId() async {
-    if (await SharedPref.getValue(PrefKey.isFirstLaunch) == false) return;
+    if (SharedPref.get(PrefKey.isFirstLaunch) == false) return;
 
     final referrer = await AndroidPlayInstallReferrer.installReferrer;
 
     final cyId = referrer.installReferrer;
 
-    if (cyId == null) return;
+    // Skip if no referrer or if it's just the default Google Play organic referrer
+    if (cyId == null || cyId == kDefaultReferrer) return;
 
-    await SharedPref.storeValue(PrefKey.cyId, cyId);
+    await SharedPref.store(PrefKey.cyId, cyId);
   }
 
   Future<void> handleDeepLinking() async {
-    if (await SharedPref.getValue(PrefKey.cyId) != null) return;
+    if (SharedPref.get(PrefKey.cyId) != null) return;
 
     final appLinks = AppLinks();
 
@@ -104,36 +97,45 @@ class InitializeService {
 
       var cyId = pathSegments.length > 1 ? pathSegments[1] : pathSegments[0];
 
-      await SharedPref.storeValue(PrefKey.cyId, cyId);
-      developer.log('Deep linking: $cyId');
+      await SharedPref.store(PrefKey.cyId, cyId);
 
       final context = navigatorKey.currentContext;
 
       if (context != null && context.mounted) {
-        await GoRouter.of(context).push(Routes.deepLinked);
+        await GoRouter.of(context).push(Routes.deepLinking);
       }
     });
   }
 
-  Future<void> initializeVersion() async {
-    final version = InfoService.instance.packageInfo.version;
+  Future<bool> initialApiCall() async {
+    try {
+      final version = InfoService.instance.packageInfo.version;
 
-    final versionDataEither = await initializeAPI.initialize(version);
+      final initialData = await initializeAPI.initialize(version);
 
-    switch (versionDataEither) {
-      case Left(value: final l):
-        developer.log(l.toString());
-      case Right(value: final r):
-        await InfoService.instance.initVersionData(r);
+      if (initialData.user != null) {
+        userController.updateCurrentUser(initialData.user!);
+        // Store doneToday from API user
+        await SharedPref.store(PrefKey.doneToday, initialData.user!.doneToday);
+        // Store last logged in email
+        await SharedPref.store(PrefKey.lastLoggedInEmail, initialData.user!.email);
+      }
+
+      await InfoService.instance.initVersionData(initialData);
+      return true;
+    } catch (e, stackTrace) {
+      developer.log('Error during initialize', error: e.toString(), stackTrace: stackTrace);
+      return false;
     }
   }
 }
 
 final initializeServiceProvider = FutureProvider<InitializeService>((ref) async {
   final service = InitializeService(
-    userController: ref.read(userControllerProvider.notifier),
+    userControllerState: ref.read(userControllerProvider),
     initializeAPI: ref.read(initializeAPIService),
-    orderedIdNotifier: ref.read(orderedIdsNotifierProvider.notifier),
+    levelController: ref.read(levelControllerProvider.notifier),
+    userController: ref.read(userControllerProvider.notifier),
   );
 
   await service.initialize();
