@@ -1,11 +1,14 @@
 import 'dart:convert';
 import 'dart:io';
+import 'package:dio/dio.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:fpdart/fpdart.dart';
 import 'package:myapp/apis/level_api.dart';
 import 'package:myapp/constants/constants.dart';
 import 'package:myapp/core/error/failure.dart';
 import 'package:myapp/core/services/file_service.dart';
+import 'package:myapp/core/services/path_service.dart';
+import 'package:myapp/core/shared_pref.dart';
 import 'package:myapp/core/utils.dart';
 import 'package:myapp/models/level/level.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
@@ -15,61 +18,25 @@ part 'level_service.g.dart';
 class LevelService {
   final FileService fileService;
   final ILevelApi levelApi;
+  final PathService pathService;
 
-  LevelService(this.fileService, this.levelApi);
+  LevelService(this.fileService, this.levelApi, this.pathService);
 
-  String get levelsDocDirPath => '${fileService.documentsDirectory.path}/levels';
-  String get levelsCacheDirPath => '${fileService.cacheDirectory.path}/levels';
-
-  String getLevelPath(String levelId) {
-    return '$levelsDocDirPath/$levelId';
-  }
-
-  String getLevelJsonPath(String levelId) {
-    return '$levelsDocDirPath/$levelId/data.json';
-  }
-
-  String getZipPath(String levelId, int zipId) {
-    return '${getZipBasePath(levelId)}/$zipId.zip';
-  }
-
-  String getZipBasePath(String levelId) {
-    return '${getLevelPath(levelId)}/zips';
-  }
-
-  String getVideoDirPath(String levelId) {
-    return '$levelsCacheDirPath/videos/$levelId';
-  }
-
-  String getUnzippedVideoPath(String levelId, String videoId) {
-    return '${getVideoDirPath(levelId)}/$videoId.mp4';
-  }
-
-  Future<Directory?> extractStoredZip(String levelId, int zipId) async {
-    final file = File(getZipPath(levelId, zipId));
-    final destinationDir = Directory(getVideoDirPath(levelId));
-    return fileService.unzip(file, destinationDir);
-  }
-
-  Future<void> deleteStoredZip(String levelId, int zipId) async {
-    await fileService.deleteFile(getZipPath(levelId, zipId));
-  }
-
-  Future<bool> isZipExists(String levelId, int zipId) async {
-    final file = File(getZipPath(levelId, zipId));
+  Future<bool> videoExists(String levelId, String videoFilename) async {
+    final file = File(pathService.fullVideoLocalPath(levelId, videoFilename));
     return file.exists();
   }
 
   Future<void> _saveLevel(LevelDTO level) async {
-    final file = File(getLevelJsonPath(level.id));
+    final file = File(pathService.levelJsonFullPath(level.id));
 
     await file.parent.create(recursive: true);
 
     await file.writeAsString(jsonEncode(level.toJson()));
   }
 
-  Future<LevelDTO?> _getLevel(String levelId) async {
-    final file = File(getLevelJsonPath(levelId));
+  Future<LevelDTO?> getLocalLevel(String levelId) async {
+    final file = File(pathService.levelJsonFullPath(levelId));
 
     if (!await file.exists()) return null;
 
@@ -82,22 +49,31 @@ class LevelService {
 
   FutureEither<LevelDTO> getLevel(String id) async {
     try {
-      final levelEither = await levelApi.getById(id);
+      final level = await levelApi.get(id);
 
-      return switch (levelEither) {
-        Right(value: final r) => r == null
-            ? await _getLevel(id).then((level) {
-                if (level == null) {
-                  return left(Failure(message: unknownErrorMsg, trace: StackTrace.current));
-                }
-                return right(level);
-              })
-            : _saveLevel(r).then(
-                (b) => right(r),
-              ),
-        Left(value: final l) => left(l),
-      };
+      if (level != null) {
+        await _saveLevel(level);
+
+        return right(level);
+      }
+
+      final localLevel = await getLocalLevel(id);
+
+      if (localLevel == null) {
+        return left(
+          Failure(
+            message: connectionErrorMsg,
+            trace: StackTrace.current,
+            type: DioExceptionType.connectionError,
+          ),
+        );
+      }
+
+      return right(localLevel);
     } catch (e, st) {
+      //remove eTag from shared pref
+      await SharedPref.removeValue(PrefKey.eTag(pathService.levelJsonPath(id)));
+
       return left(Failure(message: e.toString(), trace: st));
     }
   }
@@ -107,5 +83,6 @@ class LevelService {
 LevelService levelService(Ref ref) {
   final fileService = ref.read(fileServiceProvider);
   final levelApi = ref.read(levelApiProvider);
-  return LevelService(fileService, levelApi);
+  final pathService = ref.read(pathServiceProvider);
+  return LevelService(fileService, levelApi, pathService);
 }
