@@ -1,121 +1,66 @@
+import 'dart:developer' as developer;
 import 'dart:io';
+import 'package:flutter/foundation.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:fpdart/fpdart.dart';
 import 'package:myapp/apis/sub_level_api.dart';
+import 'package:myapp/core/services/api_service.dart';
 import 'package:myapp/core/services/level_service.dart';
+import 'package:myapp/core/services/path_service.dart';
 import 'package:myapp/core/shared_pref.dart';
-import 'package:myapp/core/utils.dart';
 import 'package:myapp/models/level/level.dart';
 
 class SubLevelService {
   final LevelService levelService;
+  final PathService pathService;
 
-  SubLevelService(this.subLevelAPI, this.levelService);
+  SubLevelService(this.subLevelAPI, this.levelService, this.pathService);
 
   final ISubLevelAPI subLevelAPI;
 
-  Future<void> getZipFiles(
-    LevelDTO levelDTO,
-  ) async {
-    final zipNums = levelDTO.sub_levels.map((subLevelDTO) => subLevelDTO.zipNum).toSet();
-
-    // First fetch all zip files in parallel
+  Future<void> getVideoFiles(LevelDTO levelDTO) async {
+    // First fetch all video files in parallel
     await Future.wait(
-      zipNums.map(
-        (zipNum) => getZip(levelDTO.id, zipNum),
+      levelDTO.sub_levels.map(
+        (subLevelDTO) => downloadVideo(levelDTO.id, subLevelDTO.videoFilename),
       ),
     );
-
-    // Then extract all zips in parallel and handle failures
-    await Future.wait(
-      zipNums.map((zipNum) async {
-        final unzipDir = await levelService.extractStoredZip(
-          levelDTO.id,
-          zipNum,
-        );
-
-        if (unzipDir == null) {
-          await SharedPref.removeRawValue(
-            PrefKey.eTagKey(
-              getLevelZipPath(
-                levelDTO.id,
-                zipNum,
-              ),
-            ),
-          );
-        }
-
-        return unzipDir;
-      }),
-    );
   }
 
-  Future<String?> getSubLevelFile(String levelId, int zipNum) async {
-    await getZip(levelId, zipNum);
+  Future<void> getSubLevelFile(String levelId, String videoFilename) async {
+    await downloadVideo(levelId, videoFilename);
 
-    final unzipDir = await levelService.extractStoredZip(levelId, zipNum);
-
-    if (unzipDir == null) {
-      SharedPref.removeRawValue(
-        PrefKey.eTagKey(
-          getLevelZipPath(
-            levelId,
-            zipNum,
-          ),
-        ),
-      );
+    if (!await levelService.videoExists(levelId, videoFilename)) {
+      SharedPref.removeValue(PrefKey.eTag(pathService.videoPath(levelId, videoFilename)));
     }
-
-    return unzipDir?.path;
   }
 
-  Future<String?> getZip(String levelId, int zipNum) async {
-    final zipDataEither = await subLevelAPI.getZipData(
-      levelId,
-      zipNum,
-    );
+  Future<void> downloadVideo(String levelId, String videoFilename) async {
+    try {
+      final videoData = await subLevelAPI.getVideo(levelId, videoFilename);
 
-    return switch (zipDataEither) {
-      Left() => _getExistingZipPath(
-          levelId,
-          zipNum,
-        ),
-      Right(
-        value: final zipData,
-      )
-          when zipData == null =>
-        _getExistingZipPath(
-          levelId,
-          zipNum,
-        ),
-      Right(
-        value: final zipData,
-      ) =>
-        _storeZipFile(
-          levelId,
-          zipNum,
-          zipData!,
-        ),
-    };
+      if (videoData == null) return;
+
+      await _storeVideo(levelId, videoFilename, videoData);
+    } catch (e) {
+      developer.log('Error downloading video: $e');
+    }
   }
 
-  String _getExistingZipPath(String levelId, int zipNum) {
-    return levelService.getZipPath(levelId, zipNum);
-  }
-
-  Future<String> _storeZipFile(String levelId, int zipNum, List<int> zipData) async {
-    final file = File(levelService.getZipPath(levelId, zipNum));
+  Future<void> _storeVideo(String levelId, String videoFilename, Uint8List videoData) async {
+    final file = File(pathService.fullVideoLocalPath(levelId, videoFilename));
     await Directory(file.parent.path).create(recursive: true);
+    await file.writeAsBytes(videoData);
+  }
 
-    await file.writeAsBytes(zipData);
-
-    return file.path;
+  String getVideoUrl(String levelId, String videoFilename, BaseUrl baseUrl) {
+    return '${baseUrl.url}${pathService.videoPath(levelId, videoFilename)}';
   }
 }
 
 final subLevelServiceProvider = Provider<SubLevelService>((ref) {
   final subLevelAPI = ref.watch(subLevelAPIProvider);
   final levelService = ref.watch(levelServiceProvider);
+  final pathService = ref.watch(pathServiceProvider);
 
-  return SubLevelService(subLevelAPI, levelService);
+  return SubLevelService(subLevelAPI, levelService, pathService);
 });
