@@ -1,10 +1,23 @@
 import 'package:flutter/material.dart';
 
 class Loader extends StatelessWidget {
-  const Loader({super.key});
+  final String? text;
+
+  const Loader({super.key, this.text});
 
   @override
   Widget build(BuildContext context) {
-    return const Center(child: CircularProgressIndicator());
+    return Center(
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          const CircularProgressIndicator.adaptive(),
+          if (text != null) ...[
+            const SizedBox(height: 16),
+            Text(text!, style: Theme.of(context).textTheme.bodyMedium),
+          ],
+        ],
+      ),
+    );
   }
 }
