@@ -11,32 +11,16 @@ Future<void> showConfirmationDialog(
     context: context,
     builder: (context) {
       return AlertDialog(
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(16),
-        ),
-        title: const Text(
-          'Confirm',
-          style: TextStyle(
-            fontWeight: FontWeight.bold,
-            fontSize: 20,
-          ),
-        ),
-        content: Text(
-          question,
-          style: const TextStyle(
-            fontSize: 16,
-          ),
-        ),
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+        title: const Text('Confirm', style: TextStyle(fontWeight: FontWeight.bold, fontSize: 20)),
+        content: Text(question, style: const TextStyle(fontSize: 16)),
         actions: [
           TextButton(
             onPressed: () {
               Navigator.of(context).pop();
               onResult(false);
             },
-            style: noButtonStyle ??
-                TextButton.styleFrom(
-                  foregroundColor: Colors.grey,
-                ),
+            style: noButtonStyle ?? TextButton.styleFrom(foregroundColor: Colors.grey),
             child: const Text('No'),
           ),
           ElevatedButton(
@@ -44,12 +28,11 @@ Future<void> showConfirmationDialog(
               Navigator.of(context).pop();
               onResult(true);
             },
-            style: yesButtonStyle ??
+            style:
+                yesButtonStyle ??
                 ElevatedButton.styleFrom(
                   backgroundColor: Colors.redAccent,
-                  shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(12),
-                  ),
+                  shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
                 ),
             child: const Text('Yes'),
           ),
