@@ -7,10 +7,7 @@ import 'package:myapp/core/widgets/loader.dart';
 class VersionCheckWrapper extends ConsumerStatefulWidget {
   final Widget child;
 
-  const VersionCheckWrapper({
-    super.key,
-    required this.child,
-  });
+  const VersionCheckWrapper({super.key, required this.child});
 
   @override
   ConsumerState<VersionCheckWrapper> createState() => _VersionCheckWrapperState();
@@ -47,20 +44,21 @@ class _VersionCheckWrapperState extends ConsumerState<VersionCheckWrapper> {
     }
 
     return Scaffold(
-      appBar: closable
-          ? AppBar(
-              backgroundColor: Colors.transparent,
-              elevation: 0,
-              actions: [
-                IconButton(
-                  icon: const Icon(Icons.close),
-                  onPressed: () {
-                    ref.read(versionControllerProvider.notifier).dismissMessage();
-                  },
-                ),
-              ],
-            )
-          : null,
+      appBar:
+          closable
+              ? AppBar(
+                backgroundColor: Colors.transparent,
+                elevation: 0,
+                actions: [
+                  IconButton(
+                    icon: const Icon(Icons.close),
+                    onPressed: () {
+                      ref.read(versionControllerProvider.notifier).dismissMessage();
+                    },
+                  ),
+                ],
+              )
+              : null,
       body: SafeArea(
         child: Center(
           child: HtmlWidget(
