@@ -5,16 +5,10 @@ import 'package:myapp/models/activity_log/activity_log.dart';
 class ActivityLogControllerState {
   final bool loading;
 
-  ActivityLogControllerState({
-    this.loading = false,
-  });
+  ActivityLogControllerState({this.loading = false});
 
-  ActivityLogControllerState copyWith({
-    bool? loading,
-  }) {
-    return ActivityLogControllerState(
-      loading: loading ?? this.loading,
-    );
+  ActivityLogControllerState copyWith({bool? loading}) {
+    return ActivityLogControllerState(loading: loading ?? this.loading);
   }
 }
 
@@ -32,6 +26,6 @@ class ActivityLogController extends StateNotifier<ActivityLogControllerState> {
 
 final activityLogControllerProvider =
     StateNotifierProvider<ActivityLogController, ActivityLogControllerState>((ref) {
-  final activityLogAPI = ref.read(activityLogAPIProvider);
-  return ActivityLogController(activityLogAPI);
-});
+      final activityLogAPI = ref.read(activityLogAPIProvider);
+      return ActivityLogController(activityLogAPI);
+    });
