@@ -1,5 +1,3 @@
-import 'dart:developer';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
@@ -8,8 +6,8 @@ import 'package:myapp/core/shared_pref.dart';
 import 'package:myapp/core/util_types/progress.dart';
 import 'package:myapp/core/widgets/loader.dart';
 import 'package:myapp/core/widgets/show_confirmation_dialog.dart';
-import 'package:myapp/features/sublevel/sublevel_controller.dart';
 import 'package:myapp/features/user/user_controller.dart';
+import 'package:myapp/models/models.dart';
 import '../auth_controller.dart';
 import '../../../core/widgets/custom_app_bar.dart';
 
@@ -20,12 +18,8 @@ class SignInScreen extends ConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     final isLoading = ref.watch(authControllerProvider.select((state) => state.loading));
 
-    log('sign in screen', name: 'sign in screen');
-
     return Scaffold(
-      appBar: const CustomAppBar(
-        title: 'Sign In',
-      ),
+      appBar: CustomAppBar(title: 'Sign In', ignoreInteractions: isLoading),
       body: Center(
         child: Padding(
           padding: const EdgeInsets.only(bottom: 32.0),
@@ -51,10 +45,7 @@ class SignInScreen extends ConsumerWidget {
                         ),
                         Text(
                           "CodeYogi's English Course",
-                          style: TextStyle(
-                            fontSize: 24,
-                            fontWeight: FontWeight.bold,
-                          ),
+                          style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
                           textAlign: TextAlign.center,
                         ),
                       ],
@@ -62,21 +53,22 @@ class SignInScreen extends ConsumerWidget {
                   ),
                   const SizedBox(height: 32),
                   ElevatedButton.icon(
-                    onPressed: isLoading
-                        ? null
-                        : () async {
-                            await ref
-                                .read(authControllerProvider.notifier)
-                                .signInWithGoogle(context);
+                    onPressed:
+                        isLoading
+                            ? null
+                            : () async {
+                              await ref
+                                  .read(authControllerProvider.notifier)
+                                  .signInWithGoogle(context);
 
-                            if (!context.mounted) return;
+                              if (!context.mounted) return;
 
-                            final user = ref.read(userControllerProvider).currentUser;
+                              final user = ref.read(userControllerProvider).currentUser;
 
-                            if (user != null) {
-                              context.go(Routes.home);
-                            }
-                          },
+                              if (user != null) {
+                                context.go(Routes.home);
+                              }
+                            },
                     icon: Image.network(
                       'https://cdn1.iconfinder.com/data/icons/google-s-logo/150/Google_Icons-09-512.png',
                       height: 24,
@@ -101,32 +93,24 @@ class SignInScreen extends ConsumerWidget {
   }
 }
 
-showLevelChangeConfirmationDialog(
-    BuildContext context, int maxLevel, int maxSubLevel, SublevelController sublevelController) {
+showLevelChangeConfirmationDialog(BuildContext context, UserModel user) {
   return showConfirmationDialog(
     context,
     question:
-        'We notice that you are already at Level $maxLevel, Sublevel $maxSubLevel. Do you want to continue from there?',
+        'We notice that you are already at Level ${user.maxLevel}, Sublevel ${user.maxSubLevel}. Do you want to continue from there?',
     onResult: (result) async {
-      if (result) {
-        await SharedPref.copyWith(
-          PrefKey.currProgress,
-          Progress(
-            level: maxLevel,
-            subLevel: maxSubLevel,
-          ),
-        );
-      }
+      final guestProgress = SharedPref.get(PrefKey.currProgress(userEmail: null)) ?? Progress();
 
-      await SharedPref.copyWith(
-        PrefKey.currProgress,
-        Progress(
-          maxLevel: maxLevel,
-          maxSubLevel: maxSubLevel,
+      await SharedPref.store(
+        PrefKey.currProgress(userEmail: user.email),
+        guestProgress.copyWith(
+          level: result ? user.maxLevel : null,
+          subLevel: result ? user.maxSubLevel : null,
+          maxLevel: user.maxLevel,
+          maxSubLevel: user.maxSubLevel,
+          levelId: result ? user.levelId : null,
         ),
       );
-
-      await sublevelController.handleFetchSublevels();
     },
     yesButtonStyle: ElevatedButton.styleFrom(
       backgroundColor: Colors.blue,
