@@ -1,12 +1,14 @@
 import 'dart:developer' as developer;
 
 import 'package:flutter/material.dart';
+import 'package:myapp/core/widgets/handle_permission_cancel.dart';
+import 'package:permission_handler/permission_handler.dart';
 import 'package:speech_to_text/speech_recognition_result.dart';
 import 'package:speech_to_text/speech_to_text.dart' as stt;
 
 class SpeechRecognizer {
   final Function(SpeechRecognitionResult) onResult;
-  final VoidCallback? onStopListenting;
+  final VoidCallback? onStopListening;
   final Function(String)? onError;
   final Function(String) onStatusChange;
 
@@ -16,36 +18,45 @@ class SpeechRecognizer {
     required this.onStatusChange,
     required this.onResult,
     this.onError,
-    this.onStopListenting,
+    this.onStopListening,
   }) {
     _speech = stt.SpeechToText();
   }
 
-  Future<void> startListening() async {
+  Future<void> _showMicPermissionDeniedDialog(BuildContext context) async {
+    await handlePermissionDenied(
+      context,
+      'Microphone permission is denied. Please open app settings and grant permission to use this feature.',
+      permission: Permission.microphone,
+    );
+  }
+
+  Future<void> startListening(BuildContext context) async {
     bool available = await _speech.initialize(
       onStatus: onStatusChange,
       onError: (error) => {if (onError != null) onError!(error.errorMsg)},
     );
 
-    if (available) {
-      _speech.listen(
-        pauseFor: const Duration(minutes: 1),
-        listenFor: const Duration(minutes: 1),
-        onResult: onResult,
-        listenOptions: stt.SpeechListenOptions(
-          partialResults: true,
-          listenMode: stt.ListenMode.dictation,
-        ),
-      );
-    } else {
-      if (onError != null) onError!('Speech recognition is not available!');
+    if (!available && context.mounted) {
+      _showMicPermissionDeniedDialog(context);
+      return;
     }
+
+    _speech.listen(
+      pauseFor: const Duration(minutes: 1),
+      listenFor: const Duration(minutes: 1),
+      onResult: onResult,
+      listenOptions: stt.SpeechListenOptions(
+        partialResults: true,
+        listenMode: stt.ListenMode.dictation,
+      ),
+    );
   }
 
   Future<void> stopListening() async {
     await _speech.stop();
     try {
-      if (onStopListenting != null) onStopListenting!();
+      if (onStopListening != null) onStopListening!();
     } catch (e) {
       developer.log('error in speech exercise widget $e');
     }
