@@ -1,7 +1,8 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:myapp/apis/user_api.dart';
 import 'package:myapp/core/shared_pref.dart';
-import 'package:myapp/features/sublevel/ordered_ids_notifier.dart';
+import 'package:myapp/core/util_types/progress.dart';
+import 'package:myapp/features/sublevel/level_controller.dart';
 import 'package:myapp/models/models.dart';
 import 'dart:developer' as developer;
 
@@ -12,30 +13,27 @@ part 'user_controller.g.dart';
 
 @freezed
 class UserControllerState with _$UserControllerState {
-  const factory UserControllerState({
-    @Default(false) bool loading,
-    UserModel? currentUser,
-  }) = _UserControllerState;
+  const factory UserControllerState({@Default(false) bool loading, UserModel? currentUser}) =
+      _UserControllerState;
 
   const UserControllerState._();
 
-  Future<int> get level async {
-    final progress = await SharedPref.getValue(PrefKey.currProgress);
+  Progress? get progress => SharedPref.get(PrefKey.currProgress(userEmail: currentUser?.email));
+
+  int get level {
     return progress?.level ?? currentUser?.level ?? 1;
   }
 
-  Future<int> get subLevel async {
-    final progress = await SharedPref.getValue(PrefKey.currProgress);
+  int get subLevel {
     return progress?.subLevel ?? currentUser?.subLevel ?? 1;
   }
 
-  Future<String?> get levelId async {
-    final progress = await SharedPref.getValue(PrefKey.currProgress);
+  String? get levelId {
     return progress?.levelId ?? currentUser?.levelId;
   }
 }
 
-@riverpod
+@Riverpod(keepAlive: true)
 class UserController extends _$UserController {
   late final IUserAPI _userAPI;
 
@@ -45,47 +43,37 @@ class UserController extends _$UserController {
     return const UserControllerState();
   }
 
-  Future<UserModel?> getCurrentUser() async {
-    state = state.copyWith(loading: true);
-
-    try {
-      final authToken = await SharedPref.getValue(PrefKey.googleIdToken);
-      if (authToken == null) return null;
-
-      final userDTO = await _userAPI.getUser();
-      if (userDTO == null) return null;
-
-      return updateCurrentUser(userDTO);
-    } catch (e, stack) {
-      developer.log('Error in getCurrentUser', error: e, stackTrace: stack);
-      return null;
-    } finally {
-      state = state.copyWith(loading: false);
-    }
-  }
-
   UserModel updateCurrentUser(UserDTO userDTO) {
-    final orderedIds = ref.read(orderedIdsNotifierProvider).value;
-    final userMaxLevel = (orderedIds?.indexOf(userDTO.maxLevelId) ?? 0) + 1;
-    final userLevel = (orderedIds?.indexOf(userDTO.levelId) ?? 0) + 1;
+    final orderedIds = ref.read(levelControllerProvider).value;
+
+    final maxLevelIndex = orderedIds?.indexOf(userDTO.maxLevelId) ?? -1;
+    final userMaxLevel = maxLevelIndex != -1 ? maxLevelIndex + 1 : 1;
+
+    final levelIndex = orderedIds?.indexOf(userDTO.levelId) ?? -1;
+    final userLevel = levelIndex != -1 ? levelIndex + 1 : 1;
 
     final user = UserModel.fromUserDTO(userDTO, userLevel, userMaxLevel);
+
     state = state.copyWith(currentUser: user);
+
     return user;
   }
 
   void removeCurrentUser() {
     state = state.copyWith(currentUser: null);
+
+    SharedPref.removeValue(PrefKey.lastLoggedInEmail);
   }
 
-  Future<void> sync(String levelId, int subLevel) async {
+  Future<bool> sync(String levelId, int subLevel) async {
     try {
       final user = await _userAPI.sync(levelId, subLevel);
-      if (user != null) {
-        updateCurrentUser(user);
-      }
+
+      updateCurrentUser(user);
+      return true;
     } catch (e, stack) {
       developer.log('Error in sync', error: e, stackTrace: stack);
+      return false;
     }
   }
 }
